questao02/requirements.txt
Scrapy==2.8.0
PyJWT==1.7.1
python-dotenv==1.0.0